+++ src/main/java/com/sg/silvergarden/service/member/MemberService.java
@@ -70,4 +70,29 @@ public class MemberService{
         result = memberDao.counselInsert(cMap);
         return result;
     }
+
+    public List<Map<String, Object>> shuttleList(Map<String, Object> sMap) {
+        List<Map<String,Object>> sList = null;
+        sList=memberDao.shuttleList(sMap);
+        return sList;
+    }
+
+    public int shuttleDelete(Map<String, Object> sMap) {
+        int result = 0;
+        result = memberDao.shuttleDelete(sMap);
+        return result;
+
+    }
+    public int shuttleUpdate(Map<String, Object> sMap) {
+        int result = 0;
+        result = memberDao.shuttleUpdate(sMap);
+        return result;
+
+    }
+    public int shuttleInsert(Map<String, Object> sMap) {
+        int result = 0;
+        result = memberDao.shuttleInsert(sMap);
+        //  }
+        return result;
+    }
 }
